generalized the result to be usable with other tools and put it in appropriate place, renamed CustomShovelInteract to CustomShovelBlockInteract to be more accurate with what it does.
