+++ gradle.properties
@@ -9,7 +9,7 @@ yarn_mappings=1.20+build.1
 loader_version=0.14.22
 
 # Mod Properties
-mod_version=1.0.1
+mod_version=1.2.0
 mod_minecraft_base_version=1.20.X
 
 # Maven/Publishing
