+++ src/main/java/net/rotgruengelb/quirl/behavior/shovel/v1/CustomShovelBlockInteract.java
@@ -3,16 +3,16 @@ package net.rotgruengelb.quirl.behavior.shovel.v1;
 import net.minecraft.block.BlockState;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
-import net.rotgruengelb.quirl.behavior.shovel.v1.result.ShovelInteractResult;
+import net.rotgruengelb.quirl.behavior.result.v1.ToolBlockInteractResult;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class CustomShovelInteract {
+public class CustomShovelBlockInteract {
 
-    public static Map<BlockState, ShovelInteractResult> CUSTOM_SHOVEL_INTERACT_RESULTS = new HashMap<>();
+    public static Map<BlockState, ToolBlockInteractResult> CUSTOM_SHOVEL_INTERACT_RESULTS = new HashMap<>();
 
     public static void addResult(BlockState starting_blockstate, BlockState new_blockstate, SoundEvent interact_sound_event, SoundCategory interact_sound_category) {
-        CUSTOM_SHOVEL_INTERACT_RESULTS.put(starting_blockstate, new ShovelInteractResult(new_blockstate, interact_sound_event, interact_sound_category));
+        CUSTOM_SHOVEL_INTERACT_RESULTS.put(starting_blockstate, new ToolBlockInteractResult(new_blockstate, interact_sound_event, interact_sound_category));
     }
 }
