+++ src/main/java/net/rotgruengelb/quirl/mixin/ShovelItemMixin.java
@@ -10,25 +10,25 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 import net.minecraft.world.event.GameEvent;
-import net.rotgruengelb.quirl.behavior.shovel.v1.result.ShovelInteractResult;
+import net.rotgruengelb.quirl.behavior.result.v1.ToolBlockInteractResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import static net.rotgruengelb.quirl.behavior.shovel.v1.CustomShovelInteract.CUSTOM_SHOVEL_INTERACT_RESULTS;
+import static net.rotgruengelb.quirl.behavior.shovel.v1.CustomShovelBlockInteract.CUSTOM_SHOVEL_INTERACT_RESULTS;
 
 @Mixin(ShovelItem.class)
 public class ShovelItemMixin {
 
     @Inject(at = @At(value = "HEAD", target = "Lnet/minecraft/item/Item;useOnBlock(Lnet/minecraft/item/ItemUsageContext;)Lnet/minecraft/util/ActionResult;"), method = "useOnBlock", cancellable = true)
-    private void extended_useOn(ItemUsageContext context, CallbackInfoReturnable<ActionResult> cir) {
+    private void add_custom_useOn(ItemUsageContext context, CallbackInfoReturnable<ActionResult> cir) {
         World world = context.getWorld();
         BlockPos blockPos = context.getBlockPos();
         BlockState blockState = world.getBlockState(blockPos);
         if (context.getSide() != Direction.DOWN) {
             PlayerEntity playerEntity = context.getPlayer();
-            ShovelInteractResult shovelInteractResult = CUSTOM_SHOVEL_INTERACT_RESULTS.get(blockState);
+            ToolBlockInteractResult shovelInteractResult = CUSTOM_SHOVEL_INTERACT_RESULTS.get(blockState);
             BlockState blockState3 = null;
             if (shovelInteractResult != null && world.getBlockState(blockPos.up()).isAir()) {
                 world.playSound(playerEntity, blockPos, shovelInteractResult.getInteract_sound_event(), shovelInteractResult.getInteract_sound_category(), 1.0f, 1.0f);
